% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nameage.R
\name{plot_nameage}
\alias{plot_nameage}
\title{plot_nameage}
\usage{
plot_nameage(names, base_year = 2015, age_range, type = c("age", "year"),
  alive_geom = c("area", "bar"), facet_scales = c("free_y", "free_x",
  "free", "fixed"), fill_color = "#008b8b", line_color = "black")
}
\arguments{
\item{names}{First names as a character vector. Names are case insensitive.}

\item{base_year}{Year that the age is calculated as of.}

\item{age_range}{Limit the range of possible ages that the name could come from.
This could be useful if you know, for example, that the name is of an adult.
If missing, all ages will be considered.}

\item{type}{Whether to plot age or year on the x-axis.}

\item{alive_geom}{Choice of whether to plot as an area or bar chart.}

\item{facet_scales}{Passed to the \code{scales} argument of \code{facet_wrap}.}

\item{fill_color}{The color that the number alive gets filled in as.}

\item{line_color}{The color that the number born is colored.}
}
\value{
A \code{ggplot2} object. It will have one facet per unique name.
}
\description{
Plot the distribution of the age for a person based on name.
}
\examples{
plot_nameage(c("Anna", "Joseph"), type = "age")
plot_nameage(c("Anna", "Joseph"), type = "year")

}

